+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.1.0] - 2025-10-21
+
+- Added
+  - New CLI options: `--proactive`, `--add-toc-title`,
+    `--add-horizontal-rules`, and `--toc-title`
+- Full diff
+  - https://github.com/jsh9/markdown-toc-creator/compare/0.0.11...0.1.0
+
 ## [0.0.11] - 2025-10-12
 
 - Fixed